refactor(navbar): change file path location
